Add files via upload
This is simple project using core java and collections for a shop Inventory, it is  a menu based program that displays options to the user where the user can add any cloth item and its price through keyboard and later on the user can purchase it. It uses the concept of Arraylist to store the item and its price.